Better componentize election ui. Organized files.
