+++ packages/webapp/src/elections/components/index.ts
@@ -1,5 +1,4 @@
 export * from "./avatars";
-export * from "./pie-mer";
 export * from "./election-member-chips";
 export * from "./error-loading-election";
 export * from "./ongoing-election";
