+++ packages/webapp/src/elections/components/ongoing-election-components/index.ts
@@ -1,6 +1,6 @@
 export * from "./chiefs-round-segment";
 export * from "./completed-round-segment";
-export * from "./consensometer";
 export * from "./ongoing-round-segment";
 export * from "./round-header";
 export * from "./support-segment";
+export * from "./video-upload-button";
